requirements.txt
matplotlib==3.8.3
networkx==3.2.1
numpy==1.26.4
pandas==2.2.1
PySCIPOpt==5.0.0
PyYAML==6.0.1
PyYAML==6.0.1
scipy==1.12.0
Shapely==2.0.3
skimage==0.0
